feat: initial certmagic update work
Still a lot to do. Everywhere that `context`s are used have currently
been passed `context.TODO()`. A logger should be passed to certmagic's
lock cleanup function. Most importantly, though, it doesn't seem to be
possible to map certmagic's LibDNS providers to Lego's. In the next
commit
I will try to see how much work it is to switch to using LibDNS instead,
 which will work natively with Certmagic. The plugin registration will
 need to change, and constructors may need to be written for each
 provider. Best-case, there should be minimal changes to the
 configuration and documentation.